requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
billiard==4.2.0
blinker==1.8.2
boto3==1.34.128
botocore==1.34.128
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==4.4.2
distro==1.9.0
dnspython==2.6.1
elevenlabs==1.3.1
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
MarkupSafe==2.1.5
moviepy==1.0.3
numpy==2.0.0
openai==1.34.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
proglog==0.1.10
prompt_toolkit==3.0.47
pydantic==2.7.4
pydantic_core==2.18.4
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.6
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
